Reorganize clang interface and extract LLVM code
